proto: extract framing into session::frame
